Feat: CoreData transportMode, notes 삭제, CoreData Manager 비동기화, ScheduleList MVVM
