+++ SeoulMate/SeoulMate/Source/Modules/ScheduleList/TripItem/TripItem.swift
@@ -1,0 +1,13 @@
+//
+//  TripItem.swift
+//  SeoulMate
+//
+//  Created by 윤혜주 on 6/13/25.
+//
+
+import Foundation
+
+enum TripItem {
+    case sectionTitle(String)
+    case trip(Tour)
+}
